requirements.txt
Flask
pymongo
python-dotenv
openai
requests
Pillow
tensorflow
nltk
